Renamed pages to be consistent with List -> Detail naming convention.  Also added Campus selection to batch pages.
- BatchSearchValue and FinancialBatchService.Partial
- *.pdb files already in gitignore
+ New migration solely for financial batch pages
