+++ Rock/Rock.csproj
@@ -150,12 +150,10 @@
     <Compile Include="Migrations\201304162319340_AddNewFields.Designer.cs">
       <DependentUpon>201304162319340_AddNewFields.cs</DependentUpon>
     </Compile>
-    <Compile Include="Migrations\201304180054257_FinancialBatchNewFields.cs" />
-    <Compile Include="Migrations\201304180054257_FinancialBatchNewFields.Designer.cs">
-      <DependentUpon>201304180054257_FinancialBatchNewFields.cs</DependentUpon>
+    <Compile Include="Migrations\201304221822201_FinancialBatch.cs" />
+    <Compile Include="Migrations\201304221822201_FinancialBatch.Designer.cs">
+      <DependentUpon>201304221822201_FinancialBatch.cs</DependentUpon>
     </Compile>
-    <Compile Include="Model\BatchSearchValue.cs" />
-    <Compile Include="Model\FinancialBatchService.Partial.cs" />
     <Compile Include="Reporting\DataTransformComponent.cs" />
     <Compile Include="Reporting\DataTransformContainer.cs" />
     <Compile Include="Reporting\DataTransform\Person\ParentTransform.cs" />
@@ -1402,8 +1400,8 @@
     <EmbeddedResource Include="Migrations\201304162319340_AddNewFields.resx">
       <DependentUpon>201304162319340_AddNewFields.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Migrations\201304180054257_FinancialBatchNewFields.resx">
-      <DependentUpon>201304180054257_FinancialBatchNewFields.cs</DependentUpon>
+    <EmbeddedResource Include="Migrations\201304221822201_FinancialBatch.resx">
+      <DependentUpon>201304221822201_FinancialBatch.cs</DependentUpon>
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup />
